requirements.txt
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
streamlit>=1.28.0
